requirements.txt
Flask
Flask-CORS
gunicorn
mercadopago
python-dotenv
schedule
Flask-SQLAlchemy
psycopg2-binary
requests